Extract api controller, remove need for toplevel controller
