--- controllers/toplevel/index.js
@@ -1,32 +1,0 @@
-'use strict';
-const config = require('config');
-const moment = require('moment');
-
-const feedbackRoute = require('./feedback');
-const surveyRoute = require('./survey');
-
-module.exports = ({ router }) => {
-    // @TODO: Move to /api
-    // handle contrast shifter
-    router.get('/contrast/:mode', (req, res) => {
-        res.cacheControl = { maxAge: 1 };
-
-        let cookieName = config.get('cookies.contrast');
-        let duration = moment.duration(6, 'months').asMilliseconds();
-        let redirectUrl = req.query.url || '/';
-        if (req.params.mode === 'high') {
-            res.cookie(cookieName, req.params.mode, {
-                maxAge: duration,
-                httpOnly: false
-            });
-        } else {
-            res.clearCookie(cookieName);
-        }
-        res.redirect(redirectUrl);
-    });
-
-    surveyRoute.init({ router });
-    feedbackRoute.init({ router });
-
-    return router;
-};
